Create MainCode
This is a tool I made to check multiple different tables for updates. If the day on the most recent update dates are today or yesterday, it displays check marks. Each block has its own check. Meaning You can see which part may not be updating. I took out all the data from this clearly so its essentially a template now. if you get some images and hook up some databases with lucee this would work fine to see if your automatic table updates are working in one page. Simple, kinda lame, here it is.